Rename SampleData.js → seed-data.js
"Seed" is the standard database testing term for documents inserted
before tests run. Matches the action (insertMany of seed documents).
